Move GrowthGraphQl schema to common location. Also make module exports commonJS since Growth Apollo node doesnt have babel
