+++ experimental/origin-graphql/src/typeDefs/Attestations.js
@@ -1,4 +1,4 @@
-export default `
+module.exports = `
   extend type Query {
     facebookAuthURL: String
   }
