+++ experimental/origin-graphql/src/typeDefs/Common.js
@@ -1,0 +1,16 @@
+var price = `
+  type Price {
+    currency: String
+    amount: String
+  }
+`
+var pageInfo = `
+  type PageInfo {
+    endCursor: String
+    hasNextPage: Boolean!
+    hasPreviousPage: Boolean!
+    startCursor: String
+  }
+`
+
+module.exports = { pageInfo, price }
