+++ experimental/origin-graphql/src/typeDefs/Identity.js
@@ -1,4 +1,4 @@
-export default `
+module.exports = `
   extend type Query {
     userRegistry: UserRegistry
     identity(id: ID!): Identity
