+++ experimental/origin-graphql/src/typeDefs/Notifications.js
@@ -1,4 +1,4 @@
-export default `
+module.exports = `
   extend type Query {
     notifications: NotificationConnection
   }
